[MWPW-117914] Gnav Promo PR feedback
