+++ libs/blocks/global-navigation/base.css
@@ -1,7 +1,7 @@
 /* Custom properties (variables) */
 :root {
   /* Top navigation - box model */
-  --feds-height-nav: 64px;
+  --feds-height-nav: 63px;
   --feds-maxWidth-nav: 1440px;
   --feds-height-breadcrumbs: 33px;
   --feds-gutter: 8px;
