+++ libs/blocks/global-navigation/features/aside/aside.js
@@ -1,0 +1,26 @@
+import { loadBlock, decorateAutoBlock } from '../../../../utils/utils.js';
+import { toFragment, lanaLog, getAnalyticsValue } from '../../utilities/utilities.js';
+
+export default async function decorateAside({ headerElem, promoPath } = {}) {
+  const onError = () => {
+    headerElem.classList.remove('has-promo');
+    lanaLog({ message: 'Gnav Promo fragment not replaced, potential CLS' });
+    return '';
+  };
+
+  const fragLink = toFragment`<a href="${promoPath}">${promoPath}</a>`;
+  const fragTemplate = toFragment`<div>${fragLink}</div>`;
+  decorateAutoBlock(fragLink);
+  if (!fragLink.classList.contains('fragment')) return onError();
+  await loadBlock(fragLink).catch(() => onError());
+  const aside = fragTemplate.querySelector('.aside');
+  if (fragTemplate.contains(fragLink) || !aside) return onError();
+
+  aside.removeAttribute('data-block');
+  aside.setAttribute('daa-lh', 'Promo');
+  aside.querySelectorAll('a').forEach((link, index) => {
+    link.setAttribute('daa-ll', getAnalyticsValue(link.textContent, index + 1));
+  });
+
+  return aside;
+}
