+++ libs/blocks/global-navigation/global-navigation.js
@@ -4,8 +4,6 @@ import {
   getMetadata,
   loadIms,
   decorateLinks,
-  loadBlock as loadMiloBlock,
-  decorateAutoBlock,
 } from '../../utils/utils.js';
 import {
   toFragment,
@@ -487,27 +485,9 @@ class Gnav {
     const promoPath = getMetadata('gnav-promo-source');
     if (!promoPath) return this.elements.aside;
 
-    const onError = () => {
-      this.el.classList.remove('has-promo');
-      lanaLog({ message: 'Gnav Promo fragment not replaced, potential CLS' });
-      return this.elements.aside;
-    };
-
-    const fragLink = toFragment`<a href="${promoPath}">${promoPath}</a>`;
-    const fragTemplate = toFragment`<div>${fragLink}</div>`;
-    decorateAutoBlock(fragLink);
-    if (!fragLink.classList.contains('fragment')) return onError();
-    await loadMiloBlock(fragLink);
-    const aside = fragTemplate.querySelector('.aside');
-    if (fragTemplate.contains(fragLink) || !aside) return onError();
-
-    this.elements.aside = aside;
-    this.elements.aside.removeAttribute('data-block');
-    this.elements.aside.setAttribute('daa-lh', 'Promo');
-    this.elements.aside.querySelectorAll('a').forEach((link, index) => {
-      link.setAttribute('daa-ll', getAnalyticsValue(link.textContent, index + 1));
-    });
-
+    const { default: decorate } = await import('./features/aside/aside.js');
+    if (!decorate) return this.elements.aside;
+    this.elements.aside = await decorate({ headerElem: this.el, promoPath });
     return this.elements.aside;
   };
